+++ operator/internal/resources/constants.go
@@ -1,0 +1,17 @@
+// Copyright 2025 Redpanda Data, Inc.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.md
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0
+
+package resources
+
+import "sigs.k8s.io/controller-runtime/pkg/client"
+
+const (
+	defaultFieldOwner = client.FieldOwner("cluster.redpanda.com/operator")
+	generationLabel   = "cluster.redpanda.com/generation"
+)
